Protect endpoints

+++ CustomEd.Forum.Service/Controllers/ForumController.cs
@@ -5,9 +5,13 @@ using CustomEd.Shared.Data.Interfaces;
 using AutoMapper;
 using CustomEd.Shared.Response;
 using CustomEd.Forum.Service.Dto;
+using Microsoft.AspNetCore.Authorization;
+using CustomEd.Shared.JWT.Interfaces;
+using CustomEd.Shared.JWT;
 
 namespace CustomEd.Forum.Service.Controllers
 {
+    [Authorize(Policy = "MemberOnly")]
     [ApiController]
     [Route("/api/{classRoomId}/forum")]
     public class ForumController : ControllerBase
@@ -16,14 +20,20 @@ namespace CustomEd.Forum.Service.Controllers
         private readonly IGenericRepository<Model.User> _userRepository;
         private readonly IGenericRepository<Classroom> _classroomRepository;
         private readonly IMapper _mapper;
+        private readonly IJwtService _jwtService;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public ForumController(IGenericRepository<Message> messageRepository, IGenericRepository<Model.User> userRepository, IGenericRepository<Classroom> classroomRepository, IMapper mapper)
+        public ForumController(IGenericRepository<Message> messageRepository, IGenericRepository<Model.User> userRepository, IGenericRepository<Classroom> classroomRepository, IMapper mapper, IJwtService jwtService, IHttpContextAccessor httpContextAccessor)
         {
             _messageRepository = messageRepository;
             _userRepository = userRepository;
             _classroomRepository = classroomRepository;
             _mapper = mapper;
+            _jwtService = jwtService;
+            _httpContextAccessor = httpContextAccessor;
         }
+        
+       
 
         [HttpPost("send-message")]
         public async Task<ActionResult<SharedResponse<MessageDto>>> SendMessage(Guid classRoomId, [FromBody] CreateMessageDto createMessageDto)
@@ -34,7 +44,11 @@ namespace CustomEd.Forum.Service.Controllers
                 var errors = validationResult.Errors.Select(e => e.ErrorMessage).ToList();
                 return BadRequest(SharedResponse<MessageDto>.Fail(null, errors));
             }
+            var currentUserId = new IdentityProvider(_httpContextAccessor, _jwtService).GetUserId(); 
+
+            createMessageDto.SenderId = currentUserId;
             createMessageDto.ClassroomId = classRoomId;
+
             var message = _mapper.Map<Message>(createMessageDto);
             await _messageRepository.CreateAsync(message);
 
@@ -52,6 +66,15 @@ namespace CustomEd.Forum.Service.Controllers
                 var errors = validationResult.Errors.Select(e => e.ErrorMessage).ToList();
                 return BadRequest(SharedResponse<MessageDto>.Fail(null, errors));
             }
+
+            var currentUserId = new IdentityProvider(_httpContextAccessor, _jwtService).GetUserId();
+
+            if(updateMessageDto.SenderId != currentUserId)
+            {
+                return Unauthorized(SharedResponse<MessageDto>.Fail(null, new List<string> { "You are not authorized to update this message." }));
+            }
+
+            updateMessageDto.SenderId = currentUserId;
             updateMessageDto.ClassroomId = classRoomId;
             var message = _mapper.Map<Message>(updateMessageDto);
             await _messageRepository.UpdateAsync(message);
@@ -68,6 +91,17 @@ namespace CustomEd.Forum.Service.Controllers
             {
                 return NotFound(SharedResponse<MessageDto>.Fail(null, new List<string> { "Message not found." }));
             }
+
+            if(message.Classroom!.Id != classRoomId)
+            {
+                return BadRequest(SharedResponse<MessageDto>.Fail(null, new List<string> { "Message does not belong to this classroom." }));
+            }
+            var currentUserId = new IdentityProvider(_httpContextAccessor, _jwtService).GetUserId();
+            if (message.Sender.Id != currentUserId)
+            {
+                return Unauthorized(SharedResponse<MessageDto>.Fail(null, new List<string> { "You are not authorized to delete this message." }));
+            }
+
             await _messageRepository.RemoveAsync(message);
             var messageDto = _mapper.Map<MessageDto>(message);
             return Ok(SharedResponse<MessageDto>.Success(messageDto, "Message deleted successfully."));
@@ -83,9 +117,10 @@ namespace CustomEd.Forum.Service.Controllers
 
 
 
-        [HttpGet("unread-messages/{userId}")]
-        public async Task<ActionResult<SharedResponse<List<Message>>>> GetUnreadMessages(Guid classRoomId, Guid userId)
+        [HttpGet("unread-messages")]
+        public async Task<ActionResult<SharedResponse<List<Message>>>> GetUnreadMessages(Guid classRoomId)
         {
+            var userId = new IdentityProvider(_httpContextAccessor, _jwtService).GetUserId();
             var user = await _userRepository.GetAsync(userId);
             if (user == null)
             {
@@ -99,9 +134,10 @@ namespace CustomEd.Forum.Service.Controllers
 
         }
 
-        [HttpGet("unread-messages-count/{userId}")]
-        public async Task<ActionResult<SharedResponse<int>>> GetUnreadMessagesCount(Guid classRoomId, Guid userId)
+        [HttpGet("unread-messages-count")]
+        public async Task<ActionResult<SharedResponse<int>>> GetUnreadMessagesCount(Guid classRoomId)
         {
+            var userId = new IdentityProvider(_httpContextAccessor, _jwtService).GetUserId();
             var user = await _userRepository.GetAsync(userId);
             if (user == null)
             {
